Refactor AddFixedPayView to MVVM
