+++ Packages/Presentation/Sources/Screens/WorkSheet/CreateWorkEntry/AddFixedPayViewModel.swift
@@ -1,0 +1,78 @@
+// Copyright © 2025 Jonas Frey. All rights reserved.
+
+import Core
+import Domain
+import Foundation
+import Model
+import SwiftUI
+
+extension AddFixedPayView {
+    @Observable
+    public class ViewModel: ViewModelProtocol {
+        let dateRange: ClosedRange<Date>
+
+        var date = Date.now
+        var activity: String = ""
+        var payAmount: Double = 0
+        var worktimes: Binding<[WorkTime]>?
+        var editingItem: Binding<WorkTime>?
+        var zeroHoursShowing = false
+
+        var isSaveButtonDisabled: Bool {
+            payAmount == 0
+        }
+
+        private init() {
+            self.dateRange = Date().addingTimeInterval(GlobalConstants.lowestValidNegativeDateInterval) ... Date()
+        }
+
+        /// Creates a new AddWorkTimeView in either adding mode, adding a new work time item on save
+        /// - Parameter worktimes: The list of worktimes to append the new object at
+        public convenience init(worktimes: Binding<[WorkTime]>) {
+            self.init()
+            self.worktimes = worktimes
+            self.editingItem = nil
+        }
+
+        /// Creates a new AddWorkTimeView in editing mode, editing the given `editingItem`
+        /// - Parameter editingItem: The work time being edited
+        public convenience init(editingItem: Binding<WorkTime>) {
+            self.init()
+            self.worktimes = nil
+            self.editingItem = editingItem
+
+            // Pre-fill the values with the ones of the editingItem
+            let worktime = editingItem.wrappedValue
+            self.activity = worktime.activity ?? ""
+            self.date = worktime.date
+            self.payAmount = worktime.pay
+        }
+
+        func invertPayAmount() {
+            payAmount *= -1
+        }
+
+        func saveEntry() {
+            guard payAmount != 0 else {
+                zeroHoursShowing = true
+                return
+            }
+            var newItem = WorkTime(
+                date: date,
+                activity: activity.isEmpty ? nil : activity,
+                fixedPay: payAmount
+            )
+            if let worktimes {
+                worktimes.wrappedValue.append(newItem)
+            } else if let editingItem {
+                // Keep the old id
+                newItem.id = editingItem.wrappedValue.id
+                editingItem.wrappedValue = newItem
+            } else {
+                assertionFailure(
+                    "AddWorkTimeView was created with neither a list of worktimes, nor an editingItem."
+                )
+            }
+        }
+    }
+}
